[CHG] [General] Fix warnings shown by gcc in Ubuntu 64-bits [CHG] [通用] 修复在 64位 Ubuntu 中 gcc 提示的 warning

+++ AMCThreadPool.c
@@ -437,7 +437,7 @@ struct AMCThreadPool *AMCThreadPool_Create(size_t threadCount)
 			
 			callStat = pthread_create(&(threadArray[tmp].thread), NULL, _threadPoolTaskStatic, &(threadArray[tmp]));
 			if (0 == callStat) {
-				_DEBUG("Thread No.%d created", tmp);
+				_DEBUG("Thread No.%ld created", (long)tmp);
 				pool->freeCount += 1;
 				pool->threadTotal += 1;
 			}
